requirements.txt
# General
tqdm>=4.60.0

# Data Collection
datasets>=2.0.0
requests>=2.25.1
semanticscholar>=0.8.0
python-dotenv>=1.0.0

# Data Processing
pandas>=1.2.0
numpy>=1.19.0

# Audio Synthesis
pyttsx3>=2.90
soundfile>=0.13.1

# Model & Training (CUDA 12.4)
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.1.0
librosa>=0.10.0
transformers>=4.36.0

# RL & Reward
sentence-transformers>=2.7.0

